// types/nationbuilder.d.ts
export interface PersonResponse {
  person?:  Person;
  precinct?: Precinct;
}

export interface PeopleListResponse {
  next?:    string;
  prev?:    string;
  results: Person[];
}
// ...
export interface Precinct {
  id?:  number;
  code: string;
  name: string;
}

export interface ContactRequest {
  contact: Contact;
}

export interface ContactResponse {
  contact: Contact;
}

export interface Contact {
  type_id?:          any;
  method?:           string;
  author_id?:        number;
  sender_id:        number;
  person_id?:        number;
  recipient_id?:     number;
  status?:           string;
  broadcaster_id?:   number;
  note?:             string;
  created_at?:       Date;
  capital_in_cents?: number;
  contact_id?:       number;
  path_id?:          any;
  step_id?:          any;
}